+++ libcontainer/container_lseek.c
@@ -10,7 +10,7 @@
 #include "libcontainer.h"
 #include "container.h"
 
-int container_lseek(stream_FILE *_file, off_t offset, int whence)
+off_t container_lseek(stream_FILE *_file, off_t offset, int whence)
 {
 	container_FILE *file = (container_FILE*)_file;
 	long new_offset;
