+++ libext2/ext2_lseek.c
@@ -7,7 +7,7 @@
 #include "libext2.h"
 #include "ext2.h"
 
-int ext2_lseek(stream_FILE *_file, long offset, int whence)
+off_t ext2_lseek(stream_FILE *_file, off_t offset, int whence)
 {
 	ext2_FILE *file = (ext2_FILE*)_file;
 	long new_offset;
